requirements.txt
pandas
sklearn
matplotlib
nltk
stanza
wordcloud
pillow